requirements copy.txt
Flask
torch==2.2.2
torchvision==0.17.2
pillow
requests
gunicorn